+++ Traveling_Saleman/src/util/Heuristique.java
@@ -9,7 +9,7 @@ public class Heuristique {
 	 * @param firstTown
 	 * @return
 	 */
-	public static int[] heursitiquePlusProcheVoisin(double [][] matrix, int firstTown){
+	public static int[] heursitiquePlusProcheVoisin(int[][] matrix, int firstTown){
 		int[] permutation = new int[matrix.length];
 		int currTown=firstTown;
 		int minIndex=0;
@@ -22,11 +22,9 @@ public class Heuristique {
 			if(matrix[currTown][minIndex]==0)minIndex++;
 			for(int j=0; j<matrix.length; j++){
 				if(!visitedTowns[j]){
-//					System.out.println(" ------ M[c][m] = " + matrix[currTown][minIndex] + " , M[c][j] = " + matrix[currTown][j]);
 					if((matrix[currTown][j]!=0 && matrix[currTown][minIndex]>matrix[currTown][j])){
 						
 						minIndex=j;
-//						System.out.println("-- je suis dedans avec la valeur de minIndex = " + minIndex);
 					}
 				}
 			}
@@ -43,11 +41,6 @@ public class Heuristique {
 			
 		}
 		
-//		System.out.println();
-//		for(int j=0; j<permutation.length; j++){
-//			System.out.print(permutation[j]+ "  |  ");
-//		}
-//		System.out.println();
 		return permutation;
 	}
 }
